requirements.txt
APScheduler==3.7.0
Django==2.2.5
django-apscheduler==0.5.1
django-cors-headers==3.6.0
django-cors-middleware==1.5.0
django-redis==4.12.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
psycopg2==2.8.6
PyJWT==1.7.1
pytz==2021.1
redis==3.5.3
six==1.15.0
sqlparse==0.4.1
tzlocal==2.1
